requirements.txt
Logbook>=1.0.0
PyYAML>=3.12
beautifulsoup4>=4.5.1
colorama>=0.3.7
future>=0.16.1
lxml>=3.6.4
requests>=2.11.1
